added NuxtPage in app created index.vue in pages
